+++ src/main/java/game/arrangement/ArrangementStrategy.java
@@ -1,3 +1,9 @@
+package game.arrangement;
+
+import game.Dot;
+import piece.Dynasty;
+import piece.Piece;
+
 import java.util.Map;
 
 public interface ArrangementStrategy {
