+++ src/main/java/piece/Advisor.java
@@ -1,3 +1,7 @@
+package piece;
+
+import game.Dot;
+
 import java.util.List;
 import java.util.Map;
