+++ src/main/java/piece/Cannon.java
@@ -1,12 +1,16 @@
+package piece;
+
+import game.Dot;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
-public class Chariot extends Piece {
-    private static final String NAME = "차";
+public class Cannon extends Piece {
+    private static final String NAME = "포";
 
-    public Chariot(Dynasty dynasty) {
+    public Cannon(Dynasty dynasty) {
         super(dynasty);
     }
 
@@ -18,7 +22,7 @@ public class Chariot extends Piece {
         int dy = origin.getDy(destination);
 
         if (dx != 0 && dy != 0) {
-            throw new UnsupportedOperationException("[ERROR] 차가 이동할 수 있는 목적지가 아닙니다.");
+            throw new UnsupportedOperationException("[ERROR] 포가 이동할 수 있는 목적지가 아닙니다.");
         }
 
         if (dx == 0 && dy == 0) {
@@ -58,10 +62,25 @@ public class Chariot extends Piece {
             throw new UnsupportedOperationException("[ERROR] 같은 나라의 말은 공격할 수 없습니다.");
         }
 
-        return routesWithPiece.values()
+        if (destinationPiece != null && destinationPiece.getName().equals(NAME)) {
+            throw new UnsupportedOperationException("[ERROR] 포는 포를 공격할 수 없습니다.");
+        }
+
+        List<Piece> pieces = routesWithPiece.values()
                 .stream()
-                .noneMatch(Objects::nonNull);
+                .filter(Objects::nonNull)
+                .toList();
+
+        if (pieces.size() != 1) {
+            throw new UnsupportedOperationException("[ERROR] 포는 경로에 단 한개의 기물만 존재해야 합니다.");
+        }
+
+        if (pieces.getFirst().getName().equals(NAME)) {
+            throw new UnsupportedOperationException("[ERROR] 포끼리 뛰어 넘을 수 없습니다.");
+        }
+        return true;
     }
+
 
     @Override
     public String getName() {
