+++ src/main/java/piece/King.java
@@ -1,11 +1,14 @@
+package piece;
+
+import game.Dot;
+
 import java.util.List;
 import java.util.Map;
 
-public class Pawn extends Piece {
+public class King extends Piece {
+    public static final String NAME = "장";
 
-    public static final String NAME = "병";
-
-    public Pawn(Dynasty dynasty) {
+    public King(Dynasty dynasty) {
         super(dynasty);
     }
 
@@ -16,14 +19,6 @@ public class Pawn extends Piece {
 
         if (Math.abs(dx) + Math.abs(dy) != 1) {
             throw new UnsupportedOperationException("[ERROR] 병이 이동할 수 있는 목적지가 아닙니다.");
-        }
-
-        if (dynasty == Dynasty.HAN && dy > 0) {
-            throw new UnsupportedOperationException("[ERROR] 병은 뒤로 이동할 수 없습니다.");
-        }
-
-        if (dynasty == Dynasty.CHO && dy < 0) {
-            throw new UnsupportedOperationException("[ERROR] 병은 뒤로 이동할 수 없습니다.");
         }
 
         return List.of(); //ToDO 빈값 확인해주기!!
