+++ src/main/java/piece/Pawn.java
@@ -1,10 +1,15 @@
+package piece;
+
+import game.Dot;
+
 import java.util.List;
 import java.util.Map;
 
-public class King extends Piece {
-    public static final String NAME = "장";
+public class Pawn extends Piece {
 
-    public King(Dynasty dynasty) {
+    public static final String NAME = "병";
+
+    public Pawn(Dynasty dynasty) {
         super(dynasty);
     }
 
@@ -15,6 +20,14 @@ public class King extends Piece {
 
         if (Math.abs(dx) + Math.abs(dy) != 1) {
             throw new UnsupportedOperationException("[ERROR] 병이 이동할 수 있는 목적지가 아닙니다.");
+        }
+
+        if (dynasty == Dynasty.HAN && dy > 0) {
+            throw new UnsupportedOperationException("[ERROR] 병은 뒤로 이동할 수 없습니다.");
+        }
+
+        if (dynasty == Dynasty.CHO && dy < 0) {
+            throw new UnsupportedOperationException("[ERROR] 병은 뒤로 이동할 수 없습니다.");
         }
 
         return List.of(); //ToDO 빈값 확인해주기!!
